Get Hospital Admissions

Downloads hospital admissions by Hospital trust using download_trust_data and then optionally aggregates to either LTLA or UTLA level. This can be done either with the built in mapping or a user supplied mapping.

get_admissions(level = "trust", release_date = Sys.Date(), mapping, geo_names)

Arguments

level

Character string, defaulting to "trust". Defines the level of aggregation at which to return the data. Other supported options are "utla" for UTLA level admissions or "ltla" for LTLA level admissions.

release_date

Date, release date of data to download. Will automatically find the Thursday prior to the date specified.

mapping

data.frame containing trust_code, p_trust, geo_code and p_geo. Defaults to trust_utla_mapping if not supplied.

geo_names

A dataframe containing geo_code and geo_name. Used to assign meaningful to geographies.

Value

A data.frame of admissions by day either at trust, LTLA or UTLA levels.